enabling rtti and making a cast to build on linux

+++ annileen/main.cpp
@@ -11,14 +11,14 @@ struct PosColorVertex
 };
 static PosColorVertex s_cubeVertices[] =
 {
-    {-1.0f,  1.0f,  1.0f, 0xff000000 },
-    { 1.0f,  1.0f,  1.0f, 0xff0000ff },
-    {-1.0f, -1.0f,  1.0f, 0xff00ff00 },
-    { 1.0f, -1.0f,  1.0f, 0xff00ffff },
-    {-1.0f,  1.0f, -1.0f, 0xffff0000 },
-    { 1.0f,  1.0f, -1.0f, 0xffff00ff },
-    {-1.0f, -1.0f, -1.0f, 0xffffff00 },
-    { 1.0f, -1.0f, -1.0f, 0xffffffff },
+    {-1.0f,  1.0f,  1.0f, (uint8_t)0xff000000 },
+    { 1.0f,  1.0f,  1.0f, (uint8_t)0xff0000ff },
+    {-1.0f, -1.0f,  1.0f, (uint8_t)0xff00ff00 },
+    { 1.0f, -1.0f,  1.0f, (uint8_t)0xff00ffff },
+    {-1.0f,  1.0f, -1.0f, (uint8_t)0xffff0000 },
+    { 1.0f,  1.0f, -1.0f, (uint8_t)0xffff00ff },
+    {-1.0f, -1.0f, -1.0f, (uint8_t)0xffffff00 },
+    { 1.0f, -1.0f, -1.0f, (uint8_t)0xffffffff },
 };
 
 static const uint16_t s_cubeTriList[] =
